+++ lib/finitio/generation/alias_type.rb
@@ -1,0 +1,9 @@
+module Finitio
+  class AliasType
+
+    def generate_data(generator, world = nil)
+      generator.call(target, world)
+    end
+
+  end # class AliasType
+end # module Finitio