+++ lib/finitio/generation/hash_based_type.rb
@@ -1,0 +1,15 @@
+module Finitio
+  module HashBasedType
+
+    def generate_data(generator, world = nil)
+      tuple = {}
+      heading.each do |a|
+        if a.required or generator.flip_coin
+          tuple[a.name.to_sym] = generator.call(a.type, world)
+        end
+      end
+      tuple
+    end
+
+  end # class HashBasedType
+end # module Finitio