+++ lib/finitio/generation/heuristic.rb
@@ -1,0 +1,8 @@
+module Finitio
+  class Generation
+    class Heuristic
+    end # class Heuristic
+  end # class Generation
+end # module Finitio
+require_relative "heuristic/constant"
+require_relative "heuristic/random"