+++ lib/finitio/generation/sub_type.rb
@@ -1,0 +1,9 @@
+module Finitio
+  class SubType
+
+    def generate_data(generator, world = nil)
+      generator.call(super_type, world)
+    end
+
+  end # class SubType
+end # module Finitio